add move and copy function

+++ app/File.php
@@ -46,6 +46,79 @@ class File
      */
     public static function delete($path = null)
     {
-        return @unlink($path);
+        if (file_exists($path) === true) {
+            return @unlink($path);
+        }
+
+        return false;
+    }
+
+    /**
+     * Copy file
+     *
+     * @param string $path
+     * @param string $newPath
+     * @param boolean $rename
+     * @return boolean
+     */
+    public static function copy($path = null, $newPath = null, $rename = false)
+    {
+        if (file_exists($path) === true) {
+            $pathInfo = pathinfo($path);
+
+            if ($rename === false) {
+                $newFile = $newPath . $pathInfo['basename'];
+            } else {
+                $newFile = $newPath . date('YmdHis') . rand(10000, 99999) . '.' . $pathInfo['extension'];
+            }
+
+            if (static::checkDir($newPath) === true) {
+                return @copy($path, $newFile);
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * Move file
+     *
+     * @param string $path
+     * @param string $newPath
+     * @param boolean $rename
+     * @return boolean
+     */
+    public static function move($path = null, $newPath = null, $rename = false)
+    {
+        if (file_exists($path) === true) {
+            $pathInfo = pathinfo($path);
+
+            if ($rename === false) {
+                $newFile = $newPath . $pathInfo['basename'];
+            } else {
+                $newFile = $newPath . date('YmdHis') . rand(10000, 99999) . '.' . $pathInfo['extension'];
+            }
+
+            if (static::checkDir($newPath) === true) {
+                return @rename($path, $newFile);
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * Check if path exists, if not create one
+     *
+     * @param string $path
+     * @return boolean
+     */
+    public static function checkDir($path = null)
+    {
+        if (is_dir($path) === false) {
+            return @mkdir($path, 0777, true);
+        }
+
+        return true;
     }
 }
